+++ lib/devices.tsx
@@ -4,10 +4,8 @@
 import axios from 'axios'
 import { handleNetworkError } from './utils'
 import { store } from '../store'
-import { getJSONItem, loadPreference } from './storage'
+import { getJSONItem } from './storage'
 import { t } from 'i18next'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faDesktop, faHeadset } from '@fortawesome/free-solid-svg-icons'
 
 export const openShowEditPhysicalPhone = (phoneInformation: any, pinstatus: any) => {
   let phoneModel: any = {}
@@ -213,17 +211,5 @@ export const tableHeader = () => {
         </th>
       </tr>
     </thead>
-  )
-}
-
-export const titleTable = (deviceType: string) => {
-  return (
-    <div className='flex items-center space-x-2'>
-      <FontAwesomeIcon
-        icon={deviceType === 'webrtc' ? faHeadset : faDesktop}
-        className='h-4 w-4 flex justify-center text-gray-700 dark:text-gray-500'
-      />
-      <span>{deviceType === 'webrtc' ? t('Devices.Web phone') : t('Devices.PhoneLink')}</span>
-    </div>
   )
 }
