+++ pages/settings.tsx
@@ -25,8 +25,8 @@ import { useRouter } from 'next/router'
 import { Theme } from '../components/settings/Theme'
 import { MobileApp } from '../components/settings/MobileApp'
 import { CustomerCards } from '../components/settings/CustomerCards'
-import Profile from './profilePicture'
-import Devices from './devices'
+import Profile from '../components/settings/ProfilePicture'
+import Devices from '../components/settings/Devices'
 import { faOfficePhone } from '@nethesis/nethesis-solid-svg-icons'
 import { savePreference } from '../lib/storage'
 import { getSelectedSettingsPage } from '../lib/settings'
